+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2026-03-23
+
+### Added
+
+- Added `locator_list` channel event handler to support resource locator listing with filter and pagination support
+- Added `resource_mapping` channel event handler to support resource field mapping
+
+### Changed
+
+- Updated `@choiceopen/atomemo-plugin-schema` from ^0.5.4 to ^0.6.0
+
+### Fixed
+
+- `serializeFeature` now omits `locator_list` and `resource_mapping` from serialized tool definitions (these are runtime-only handler maps and should not be sent to the server)
+
 ## [0.5.4] - 2026-03-16
 
 ### Changed
@@ -378,7 +393,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `pino` from `^10.2.0` to `^10.2.1`
 - Updated `es-toolkit` from `^1.43.0` to `^1.44.0`
 
-[Unreleased]: https://github.com/choice-open/atomemo-plugin-sdk-js/compare/v0.5.4...HEAD
+[Unreleased]: https://github.com/choice-open/atomemo-plugin-sdk-js/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/choice-open/atomemo-plugin-sdk-js/compare/v0.5.4...v0.6.0
 [0.5.4]: https://github.com/choice-open/atomemo-plugin-sdk-js/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/choice-open/atomemo-plugin-sdk-js/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/choice-open/atomemo-plugin-sdk-js/compare/v0.5.1...v0.5.2